requirements.txt
Django==4.1.3
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
qrcode[pil]==7.3.1
python-barcode==0.14.0
django-cors-headers==3.13.0